List words via CLI
Opens up the dictionary to expose lists of "entries" comprised of their
index, kind, and the word itself. Then uses this new function to
retrieve and print the words via a command line subcommand. The command
can also be formatted as either csv or fixed width. The word is placed
after the kind so that it can be sorted.